Process scorm packages during ims imports

+++ app/jobs/upload_canvas_job.rb
@@ -36,7 +36,7 @@ class UploadCanvasJob < ApplicationJob
       )
     end
 
-    FileUtils.remove_entry_secure(file_path)
+    FileUtils.remove_entry_secure(file_path) if file_path.present?
   rescue StandardError => e
     scorm_course.update(import_job_status: ScormCourse::FAILED)
     raise e
